we do a little bit of cleaning up
gonna move some elements into jsx, and do some bug fixes with annotations, then next is the rating system.
